+++ examples/compute_plan/scripts/add_compute_plan.py
@@ -42,6 +42,7 @@ with zipfile.ZipFile(archive_path, 'w') as z:
 
 algo_key = client.add_algo({
     'name': 'SGD classifier death predictor',
+    'category': 'ALGO_SIMPLE',
     'file': archive_path,
     'description': os.path.join(algo_directory, 'description.md'),
     'permissions': {