+++ examples/debugging/scripts/local_debugging.py
@@ -91,6 +91,7 @@ with zipfile.ZipFile(archive_path, "w") as z:
 #  Create the algorithm archive and asset
 ALGO = {
     "name": "Titanic: Random Forest",
+    "category": "ALGO_SIMPLE",
     "description": assets_directory / "algo_random_forest" / "description.md",
     "file": current_directory / "tmp" / "algo_random_forest.zip",
     "permissions": {"public": False, "authorized_ids": []},
@@ -174,6 +175,7 @@ print("Adding algo...")
 algo_key = client.add_algo(
     {
         "name": ALGO["name"],
+        "category": ALGO["category"],
         "file": str(ALGO["file"]),
         "description": str(ALGO["description"]),
         "permissions": ALGO["permissions"],