+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
 - Internal: backend now serve paginated lists of asset. Client still returns simple lists as before.
+- [BREAKING CHANGE] All algos are now one single object with a category property.
 
 ## [0.10.0] - 2021-08-05